Render Partial Collections
<%= render @post.comments %> will now render the partial in
app/views/comments/_comment.html.erb
once for each comment that is in the @post.comments collection.
As the render method iterates over the @post.comments collection,
it assigns each comment to a local variable named the same as the partial
in this case 'comment', which is then available in the partial for us to show.